pyproject.toml
[tool.poetry]
name = "BigBox-Trainee"
version = "0.1.0"
description = ""
authors = ["felixhc <felixhumbertocarabajal5@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.1"
Pillow = "^8.4.0"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^21.1"
django-environ = "^0.7.0"
django-extensions = "^3.1.3"
ipdb = "^0.13.9"
argon2-cffi = "^21.1.0"
White-Noise = "^0.1.0"
gunicorn = "^20.1.0"
dj-database-url = "^0.5.0"
psycopg2 = "^2.9.1"
django-debug-toolbar = "^3.2.2"
ipython = "^7.29.0"
Werkzeug = "^2.0.2"
graphene-django = "^2.15.0"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8-isort = "^4.1.1"
flake8-black = "^0.2.3"
flake8-django = "^1.1.2"
flake8-simplify = "^0.14.2"
black = "^21.9b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
